fix for generic BL Touch
High-speed mode was causing a generic BL Touch to error out.

Retracting the pin wasn't fast enough, and it would throw an error that
it failed to retract after the first probe.

Removing the "high speed" command seems to make it work, and causes no
visible (to humans) slowdown.  I suspect it's on the order of
milliseconds.